Update nmap_serv_versions
This is a Bash script that automates Nmap scans on a specified IP address. First, the script will ask the user to enter an IP address. Then, it will initiate an Nmap scan that will scan all ports on the target machine using the -p- flag, which scans all ports, and the -Pn flag, which disables host discovery and speeds up the scan. The output of the scan is then saved to a temporary location, and parsed to extract only the number of the open ports, which are saved to the 'open_ports' variable in the comma-separated format used by Nmap.

After that, the script proceeds to conduct a second Nmap scan, which focuses only on the ports that were found to be open in the previous scan. The purpose of this scan is to gather the version of the services running on each of the open ports. To achieve this, the -sV flag is included in the Nmap command. By doing so, the script avoids scanning all 65 thousand+ ports, which could take an enormous amount of time, and instead focuses only on the ports that have already been identified as open.

This efficient approach of first identifying open ports and then scanning for service versions only on those ports, makes the script much faster and less resource-intensive, thus ensuring a smoother scanning experience. The resulting output is then saved to a file for further analysis. This script offers a quick and easy way to automate Nmap scans and gather valuable information about the services running on a particular machine.